python_scripts_housefire/requirements.txt
-e .
pandas
redis
hiredis
python-dotenv
black
requests
undetected-chromedriver
